+++ routes/authRoutes.js
@@ -1,0 +1,9 @@
+import express from "express"
+import { login } from "../controllers/authController";
+
+const authRouter = express.Router();
+
+// LOGIN
+authRouter.post("/login", login);
+
+export default authRouter;
