+++ server.js
@@ -2,6 +2,7 @@ import express from "express";
 import mongoose from "mongoose";
 import dotenv from "dotenv";
 import userRouter from "./routes/userRoutes"
+import authRouter from "./routes/authRoutes";
 dotenv.config();
 const app = express();
 
@@ -9,6 +10,8 @@ const port = 5000;
 app.use(express.json())
 // middlewares
 app.use("/user", userRouter);
+app.use("/auth", authRouter);
+
 
 mongoose
   .connect(
